pyproject.toml
[project]
name = "video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "moviepy>=2.2.1",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=11.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
